Extractor/requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.5
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
filelock==3.14.0
fonttools==4.58.5
fsspec==2025.5.1
ftfy==6.3.1
hf-xet==1.1.5
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inquirerpy==0.3.4
iopath==0.1.10
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mmcv==2.2.0
mmcv-full==1.7.2
mmengine==0.10.7
model-index==0.1.11
moviepy==2.2.1
mpmath==1.3.0
networkx==3.4.2
numpy<2.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.3.0
pfzy==0.3.4
pillow==11.1.0
pip==25.1.1
platformdirs==4.3.8
portalocker==3.2.0
proglog==0.1.12
prompt_toolkit==3.0.52
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2023.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.2
rich==13.4.2
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==60.2.0
six==1.17.0
supervision==0.25.1
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==1.0.16
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.65.2
transformers==4.53.1
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
wheel==0.45.1
yapf==0.43.0
